Add auth

+++ restaurant/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 class Booking(models.Model):
     name = models.CharField(max_length=255)
-    no_of_guests = models.IntegerField()
+    no_of_guests = models.SmallIntegerField()
     date = models.DateTimeField()
 
     def __str__(self):
@@ -14,7 +14,7 @@ class Booking(models.Model):
 class Menu(models.Model):
     title = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    inventory = models.IntegerField()
+    inventory = models.SmallIntegerField()
 
     def __str__(self):
         return self.title
